src: reduce includes of node_internals.h
PR-URL: https://github.com/nodejs/node/pull/25507
