+++ src/async_wrap.cc
@@ -22,7 +22,6 @@
 #include "async_wrap-inl.h"
 #include "env-inl.h"
 #include "node_errors.h"
-#include "node_internals.h"
 #include "tracing/traced_value.h"
 #include "util-inl.h"
