+++ src/connection_wrap.cc
@@ -3,7 +3,6 @@
 #include "connect_wrap.h"
 #include "env-inl.h"
 #include "pipe_wrap.h"
-#include "node_internals.h"
 #include "stream_base-inl.h"
 #include "stream_wrap.h"
 #include "tcp_wrap.h"
