+++ src/heap_utils.cc
@@ -1,5 +1,4 @@
-#include "node_internals.h"
-#include "env.h"
+#include "env-inl.h"
 
 using v8::Array;
 using v8::Boolean;
