+++ src/inspector_js_api.cc
@@ -1,7 +1,6 @@
 #include "base_object-inl.h"
 #include "inspector_agent.h"
 #include "inspector_io.h"
-#include "node_internals.h"
 #include "v8.h"
 #include "v8-inspector.h"
