+++ src/js_stream.cc
@@ -4,7 +4,6 @@
 #include "env-inl.h"
 #include "node_buffer.h"
 #include "node_errors.h"
-#include "node_internals.h"
 #include "stream_base-inl.h"
 #include "v8.h"
