+++ src/module_wrap.cc
@@ -7,7 +7,6 @@
 #include "node_errors.h"
 #include "node_url.h"
 #include "util-inl.h"
-#include "node_internals.h"
 #include "node_contextify.h"
 #include "node_watchdog.h"
