+++ src/node_dtrace.h
@@ -24,7 +24,8 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "node_internals.h"
+#include "env.h"
+#include "v8.h"
 
 extern "C" {
 /*
