+++ src/node_http2.cc
@@ -4,8 +4,8 @@
 #include "node_buffer.h"
 #include "node_http2.h"
 #include "node_http2_state.h"
-#include "node_internals.h"
 #include "node_perf.h"
+#include "util.h"
 
 #include <algorithm>
