+++ src/node_http_parser_impl.h
@@ -27,7 +27,7 @@
 
 #include "node.h"
 #include "node_buffer.h"
-#include "node_internals.h"
+#include "util.h"
 
 #include "async_wrap-inl.h"
 #include "env-inl.h"
