+++ src/node_i18n.h
@@ -24,7 +24,8 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "node_internals.h"
+#include "util.h"
+
 #include <string>
 
 #if defined(NODE_HAVE_I18N_SUPPORT)
