+++ src/node_messaging.cc
@@ -4,7 +4,6 @@
 #include "debug_utils.h"
 #include "node_buffer.h"
 #include "node_errors.h"
-#include "node_internals.h"
 #include "node_process.h"
 #include "util-inl.h"
 #include "util.h"
