+++ src/node_os.cc
@@ -19,8 +19,9 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-#include "node_internals.h"
+#include "env-inl.h"
 #include "string_bytes.h"
+#include "util.h"
 
 #include <array>
 #include <errno.h>
