+++ src/node_perf_common.h
@@ -4,6 +4,7 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "node.h"
+#include "uv.h"
 #include "v8.h"
 
 #include <algorithm>
