+++ src/node_worker.cc
@@ -1,7 +1,6 @@
 #include "node_worker.h"
 #include "debug_utils.h"
 #include "node_errors.h"
-#include "node_internals.h"
 #include "node_buffer.h"
 #include "node_perf.h"
 #include "util.h"
