+++ src/node_zlib.cc
@@ -21,7 +21,6 @@
 
 #include "node.h"
 #include "node_buffer.h"
-#include "node_internals.h"
 
 #include "async_wrap-inl.h"
 #include "env-inl.h"
@@ -58,6 +57,19 @@ using v8::Uint32Array;
 using v8::Value;
 
 namespace {
+
+// Fewer than 64 bytes per chunk is not recommended.
+// Technically it could work with as few as 8, but even 64 bytes
+// is low.  Usually a MB or more is best.
+#define Z_MIN_CHUNK 64
+#define Z_MAX_CHUNK std::numeric_limits<double>::infinity()
+#define Z_DEFAULT_CHUNK (16 * 1024)
+#define Z_MIN_MEMLEVEL 1
+#define Z_MAX_MEMLEVEL 9
+#define Z_DEFAULT_MEMLEVEL 8
+#define Z_MIN_LEVEL -1
+#define Z_MAX_LEVEL 9
+#define Z_DEFAULT_LEVEL Z_DEFAULT_COMPRESSION
 
 #define ZLIB_ERROR_CODES(V)      \
   V(Z_OK)                        \
