+++ src/pipe_wrap.cc
@@ -27,7 +27,6 @@
 #include "handle_wrap.h"
 #include "node.h"
 #include "node_buffer.h"
-#include "node_internals.h"
 #include "connect_wrap.h"
 #include "stream_base-inl.h"
 #include "stream_wrap.h"
