+++ src/spawn_sync.h
@@ -24,8 +24,9 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "node_internals.h"
 #include "node_buffer.h"
+#include "uv.h"
+#include "v8.h"
 
 namespace node {
