+++ src/stream_base.cc
@@ -4,11 +4,9 @@
 #include "node.h"
 #include "node_buffer.h"
 #include "node_errors.h"
-#include "node_internals.h"
 #include "env-inl.h"
 #include "js_stream.h"
 #include "string_bytes.h"
-#include "util.h"
 #include "util-inl.h"
 #include "v8.h"
