+++ src/string_decoder.cc
@@ -1,7 +1,7 @@
-#include "string_decoder-inl.h"
-#include "string_bytes.h"
-#include "node_internals.h"
+#include "env-inl.h"
 #include "node_buffer.h"
+#include "string_bytes.h"
+#include "string_decoder-inl.h"
 
 using v8::Array;
 using v8::Context;
