+++ src/string_search.h
@@ -7,7 +7,8 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "node_internals.h"
+#include "util.h"
+
 #include <string.h>
 #include <algorithm>
