+++ src/udp_wrap.cc
@@ -22,7 +22,6 @@
 #include "udp_wrap.h"
 #include "env-inl.h"
 #include "node_buffer.h"
-#include "node_internals.h"
 #include "handle_wrap.h"
 #include "req_wrap-inl.h"
 #include "util-inl.h"
